requirements.txt
apscheduler==3.10.1
blue==0.9.1
mysqlclient==2.1.1
pandas==2.0.1
pip-chill==1.0.3
pyyaml==6.0
sqlalchemy==2.0.13
